+++ ext.php
@@ -3,7 +3,7 @@
  *
  * Topic Cement Style. An extension for the phpBB Forum Software package.
  *
- * @copyright (c) 2018, phpBB Studio, https://www.phpbbstudio.com
+ * @copyright (c) 2020, phpBB Studio, https://www.phpbbstudio.com
  * @license GNU General Public License, version 2 (GPL-2.0)
  *
  */
@@ -13,46 +13,41 @@ namespace phpbbstudio\tcs;
 /**
  * Topic Cement Style Extension base
  *
- * It is recommended to remove this file from
- * an extension if it is not going to be used.
  */
 class ext extends \phpbb\extension\base
 {
 	/**
-	 * Check whether the extension can be enabled.
-	 * Provides meaningful(s) error message(s) and the back-link on failure.
-	 * CLI compatible.
+	 * Indicate whether or not the extension can be enabled.
 	 *
-	 * @return bool
+	 * @return bool				True if enableable, false otherwise.
+	 * @access public
 	 */
 	public function is_enableable()
 	{
-		$is_enableable = true;
-
-		$user = $this->container->get('user');
-		$user->add_lang_ext('phpbbstudio/tcs', 'ext_require');
-		$lang = $user->lang;
-
-		if ( !(phpbb_version_compare(PHPBB_VERSION, '3.2.7', '>=') && phpbb_version_compare(PHPBB_VERSION, '3.3.0@dev', '<')) )
+		if (!(phpbb_version_compare(PHPBB_VERSION, '3.2.7', '>=') && phpbb_version_compare(PHPBB_VERSION, '4.0.0@dev', '<')))
 		{
-			/**
-			 * Despite it seems wrong that's the right approach and not an error in coding
-			 * That's done in order to avoid a PHP error like
-			 * "Indirect modification of overloaded property phpbb/user::$lang has no effect"
-			 * Discussed here: https://www.phpbb.com/community/viewtopic.php?p=14724151#p14724151
-			*/
-			$lang['EXTENSION_NOT_ENABLEABLE'] .= '<br>' . $user->lang('ERROR_PHPBB_VERSION', '3.2.7', '3.3.0@dev');
-			$is_enableable = false;
+			if (phpbb_version_compare(PHPBB_VERSION, '3.3.0', '>='))
+			{
+				$language= $this->container->get('language');
+				$language->add_lang('ext_require', 'phpbbstudio/tcs');
+
+				return $language->lang('ERROR_PHPBB_VERSION', '3.2.7', '4.0.0@dev');
+			}
+			else
+			{
+				$user = $this->container->get('user');
+				$lang = $user->lang;
+
+				$user->add_lang_ext('phpbbstudio/tcs', 'ext_require');
+
+				$lang['EXTENSION_NOT_ENABLEABLE'] .= '<br>' . $user->lang('ERROR_PHPBB_VERSION', '3.2.7', '4.0.0@dev');
+
+				$user->lang = $lang;
+
+				return false;
+			}
 		}
 
-		if (!phpbb_version_compare(PHP_VERSION, '5.5', '>='))
-		{
-			$lang['EXTENSION_NOT_ENABLEABLE'] .= '<br>' . $user->lang('ERROR_PHP_VERSION', '5.5');
-			$is_enableable = false;
-		}
-
-		$user->lang = $lang;
-
-		return $is_enableable;
+		return true;
 	}
 }