+++ language/en/info_acp_tcs.php
@@ -3,7 +3,7 @@
  *
  * Topic Cement Style. An extension for the phpBB Forum Software package.
  *
- * @copyright (c) 2018, phpBB Studio, https://www.phpbbstudio.com
+ * @copyright (c) 2020, phpBB Studio, https://www.phpbbstudio.com
  * @license GNU General Public License, version 2 (GPL-2.0)
  *
  */
@@ -15,15 +15,12 @@ if (!defined('IN_PHPBB'))
 
 if (empty($lang) || !is_array($lang))
 {
-	$lang = array();
+	$lang = [];
 }
 
-$lang = array_merge($lang, array(
-	// Modules
-	'ACP_TCS_CAT'					=> 'phpBB Studio - Topic Cement Style',
-	'ACP_TCS_SETTINGS'				=> 'Settings',
-
-	// Forums
-	'ACP_FORUM_TOPIC_PRIORITY'		=> 'Topic Cement Style',
-	'ACP_FORUM_TOPIC_PRIORITY_DESC'	=> 'Whether or not we want to prioritise and style title of topics.',
-));
+/**
+ * Some characters you may want to copy&paste: ’ » “ ” …
+ */
+$lang = array_merge($lang, [
+	'ACL_CAT_PHPBB_STUDIO'	=> 'phpBB Studio',
+]);