+++ migrations/install_acp_schema.php
@@ -3,7 +3,7 @@
  *
  * Topic Cement Style. An extension for the phpBB Forum Software package.
  *
- * @copyright (c) 2018, phpBB Studio, https://www.phpbbstudio.com
+ * @copyright (c) 2020, phpBB Studio, https://www.phpbbstudio.com
  * @license GNU General Public License, version 2 (GPL-2.0)
  *
  */
@@ -13,59 +13,46 @@ namespace phpbbstudio\tcs\migrations;
 class install_acp_schema extends \phpbb\db\migration\migration
 {
 	/**
-	 * Check if the migration is effectively installed (entirely optional).
-	 *
-	 * @return bool 		True if this migration is installed, False if this migration is not installed
-	 * @access public
-	 */
+	* {@inheritdoc}
+	*/
 	public function effectively_installed()
 	{
 		return $this->db_tools->sql_column_exists($this->table_prefix . 'forums', 'forum_topic_priority');
 	}
 
 	/**
-	 * Assign migration file dependencies for this migration.
-	 *
-	 * @return array		Array of migration files
-	 * @access public
-	 * @static
-	 */
+	* {@inheritdoc}
+	*/
 	static public function depends_on()
 	{
-		return array('\phpbb\db\migration\data\v32x\v327');
+		return ['\phpbb\db\migration\data\v32x\v327'];
 	}
 
 	/**
-	 * Add the TCS extension forums schema to the database.
-	 *
-	 * @return array 		Array of table schema
-	 * @access public
-	 */
+	* {@inheritdoc}
+	*/
 	public function update_schema()
 	{
-		return array(
-			'add_columns'	=> array(
-				$this->table_prefix . 'forums'		=> array(
-					'forum_topic_priority'			=>	array('BOOL', 1),
-				),
-			),
-		);
+		return [
+			'add_columns'	=> [
+				$this->table_prefix . 'forums' => [
+					'forum_topic_priority'	=>	['BOOL', 1],
+				],
+			],
+		];
 	}
 
 	/**
-	 * Drop the TCS extension forums schema from the database.
-	 *
-	 * @return array		Array of table schema
-	 * @access public
-	 */
+	* {@inheritdoc}
+	*/
 	public function revert_schema()
 	{
-		return array(
-			'drop_columns'	=> array(
-				$this->table_prefix . 'forums'	=> array(
+		return [
+			'drop_columns'	=> [
+				$this->table_prefix . 'forums' => [
 					'forum_topic_priority',
-				),
-			),
-		);
+				],
+			],
+		];
 	}
 }